Time: 0 ms (100%), Space: 48.5 MB (45.56%) - LeetHub

+++ 3685-count-subarrays-of-length-three-with-a-condition/3685-count-subarrays-of-length-three-with-a-condition.cpp
@@ -17,7 +17,7 @@ public:
                 int thirdnum = nums[j];
 
                 int sum = firstnum + thirdnum;
-                float   res =  (float)secondnum / 2  ;
+                float   res = secondnum / 2.0  ;
                 if( res ==  sum  ) {
                     cnt++;
                 }
